Test In Gendarme is working. Test in Mono has been updated to use a registry key, as per jpobst's suggestion. (Ordinarily I wouldn't commit both of these features together, but it's only a couple of lines and is trivial.)
Still need to figure out a way to handle websites for Test in Mono--going to have
to look into XSP to figure out the best way to do that.

The .xsl/.xsd files are used by CloverleafET in order to generate reports, as
with swf-win32-runner (they're the same files). The Gendarme assemblies are
included as a known-working copy of the libraries CloverleafShared needs.