Another fairly big commit; I'll try to remember to commit more often. Lots of changes.
First, an organizational note: like the files in the root directory, the "tools"
directory contains helper apps upon which Cloverleaf requires: xming (zipped so
as not to shove 550 files into the tree), pscp, and plink (command-line ssh).

-Added a tiny quick-project, RemoteTestApplication, for working with remote app
deployment. All it does is pull Environment.OSVersion.ToString() to display. Why?
Because I wanted something to create these screenshots:

http://i27.tinypic.com/2rh0777.png - RemoteTestApplication running under .NET 2.0
on Windows.
http://i31.tinypic.com/2rcuyyr.png - RemoteTestApplication running under Mono 1.9.1
on Linux, X-forwarded to my developer workstation via Cloverleaf's RemoteAppTester.

-Added RemoteAppTester and RemoteAppSelector. I haven't documented them yet, but
they should be pretty obvious - it uses PSCP and Plink in order to upload the app
to the Linux machine and invoke it there. It does not yet do Zeroconf to find
Linux machines running CloverleafService, because...I haven't written the
service yet. D'oh.

-Added CloverleafService, which is currently still basically a skeleton. Need to
figure out how to compile Mono.Zeroconf.Providers.Avahi, which has so far not
been pleasant whatsoever. Then I'll be able to do the Zeroconf part of the
RemoteAppTester, followed by the ASP.NET server invocation code of both XSP
and Apache/mod_mono on the fly for a developer. Seeing the end of the tunnel on
this part of the project...