Created CloverleafAddin (again...my working copy grew very angry and I junked it).
Had to extend CloverleafEnvironment with CloverleafEnvironment.IsAddin in order to
allow for the divergent code paths (you have to prime the message pump, etc., with
a standalone app; with the addin, you just create a dialog). Haven't gotten to the
other options besides Test In Mono, but will this week, and figure out a solution
w.r.t. certain behaviors that currently act incorrectly for addins.